Feat: return decoded and raw response
When we send some queries to the server, it sends us a response that has more information in it, than is present in the decoded response.  `runCommand` for example sends back the output of the command, but the decoded response is an `emptyMessage`.

The very large discrepancy between the raw and the decoded, means there is significant value in the raw response data